Altruism Happiness - Ch12 - from Brethel-Haurwitz and Marsh (2014)

Happiness may not be important just for the person feeling it; happiness may also promote kind, altruistic behavior. Brethel-Haurwitz and Marsh (2014) examined this idea by collecting data on U.S. states. A Gallup poll in 2010 was used to measure each state's well-being index, a measure of mean happiness for the state's residents on a scale from 0 to 100. Next, a kidney donation database for 1999-2010 was used to figure out each state's rate (number of donations per 1 million people) of non-directed kidney donations-that's giving one kidney to a stranger, an extremely generous and altruistic thing to do!

Usage

data_altruism_happiness

Format

data_altruism_happiness

A data frame with 50 rows and 6 columns:

State

factor - State where data was collected

Abbreviation

factor - State where data was collected

Well_Being_2010

numeric - State where data was collected

Well_Being_2013

numeric - State where data was collected

Kidney_Rate, per million population

numeric - State where data was collected

WB Change 2013-2010

numeric - State where data was collected

Source

https://journals.sagepub.com/doi/full/10.1177/0956797613516148