Gender math IAT ma - Ch9 - Many Labs replications of Nosek et al. (2002)

In EOC Exercise 4 in Chapter 7 we encountered the classic study of Nosek et al. (2002), in which male and female participants completed an Implicit Association Test (IAT) that measured the extent of negative attitudes towards mathematics, compared with art. The study found that women, compared with men, tended to have more negative implicit attitudes towards mathematics. The Many Labs project repeated this study at locations around the world (Klein et al., 2014a, 2014b). Summary data for 30 of these labs are available in Gender math IAT ma. Higher scores indicate more implicit bias against mathematics. See also data_gender_math_iat for raw data from two specific sites from this replication effort.

Usage

data_gender_math_iat_ma

Format

data_gender_math_iat_ma

A data frame with 30 rows and 9 columns:

Location

factor

M Male

numeric

s Male

numeric

n Male

integer

M Female

numeric

s Female

numeric

n Female

integer

USAorNot

factor

Country

factor

Source

https://psycnet.apa.org/fulltext/2014-20922-002.html