Kardas Expt 3 - Ch07 - from Kardas and O'Brien (2018), Experiment 3

Suppose you want to change the battery in your phone, cook the perfect souffle, or perform a three-ball juggle. Just as numerous people do every day, you might search online to find a video that shows what to do. Suppose you watch such a video just once. First question: How well would you then predict you could perform the task? Second question: How well would you actually perform the task, the first time you tried? Now suppose you watch the video many times: Again consider the two questions. These questions were investigated in a series of studies by Kardas and O'Brien (2018). Let's first do some quick analyses of Kardas Experiments 3 and 4-let's call them Expt 3 and Expt 4-focusing on the effect of watching a video many times rather than once. In Expt 3, participants first watched a brief video of a person performing the moonwalk. The Low Exposure group watched the video once, the High Exposure group 20 times. Then participants predicted, on a 1 to 10 scale, how well they felt they would be able to perform the moonwalk themselves. Finally, they attempted a single performance of the moonwalk, which was videoed. These videos were rated, on the same 1 to 10 scale, by independent raters.

Usage

data_kardas_expt_3

Format

data_kardas_expt_3

A data frame with 100 rows and 3 columns:

Exposure

factor

Prediction

numeric

Performance

numeric

Source

doi:10.1177/0956797617740646