Kardas Expt 4 - Ch07 - from Kardas and O'Brien (2018), Experiment 4

Suppose you want to change the battery in your phone, cook the perfect souffle, or perform a three-ball juggle. Just as numerous people do every day, you might search online to find a video that shows what to do. Suppose you watch such a video just once. First question: How well would you then predict you could perform the task? Second question: How well would you actually perform the task, the first time you tried? Now suppose you watch the video many times: Again consider the two questions. These questions were investigated in a series of studies by Kardas and O'Brien (2018). Let's first do some quick analyses of Kardas Experiments 3 and 4-let's call them Expt 3 and Expt 4-focusing on the effect of watching a video many times rather than once. Expt 4 was conducted online with participants recruited from Amazon's Mechanical Turk, who are typically more diverse than students. The online task was based on a mirror-drawing game developed by Bob and students (Cusack et al., 2015, tiny.cc/bobmirrortrace). Participants first read a description of the game and the scoring procedure. To play, you use your computer trackpad to trace a target line, as accurately and quickly as you can. The task is tricky because you can see only a mirror image of the path you are tracing with a finger on the trackpad. A running score is displayed. The final score is the percentage match between the target line and the path you traced, so scores can range from 0 to 100

Usage

data_kardas_expt_4

Format

data_kardas_expt_4

A data frame with 270 rows and 4 columns:

Exposure

factor

Prediction

integer

Performance

integer

Time

numeric

Source

doi:10.1177/0956797617740646