StudyStrategies - Ch14 - from O'Reilly et al. (1998)

To what extent does study strategy influence learning? To investigate, psychology students were randomly assigned to three groups and asked to learn biology facts using one of three different strategies: a) Self-Explain (explaining for each fact what new knowledge is gained and how it relates to what is already known), b) Elab Interrogation (elaborative interrogation: stating for each fact why it makes sense), or c) Repetition Control (stating each fact over and over). After studying, students took a 25-point fill-the-blank test (O'Reilly et al., 1998)

Usage

data_studystrategies

Format

data_studystrategies

A data frame with 3 rows and 10 columns:

Group

factor

TestMean

numeric

TestSD

numeric

TestN

integer

PrevKnowMean

numeric

PrevKnowSD

numeric

PrevKnowN

integer

EaseUseMean

numeric

EaseUseSD

numeric

EaseUseN

integer

Source

doi:10.1006/ceps.1997.0977