Sample data for estimate_mdiff_paired: LSAT scores from Thomason (2014)

Thomason et al. (2014) reported 7 studies evaluating a promising approach to teaching critical thinking that combines argument mapping with a form of mastery learning (Mazur, 1997). Studies were conducted in the United States, Canada, and the United Kingdom, and each study used a single group of students. Students were tested on various well-established measures of critical thinking, both before (the Pretest) and after (the Posttest) training. Group sizes ranged from 7 to 39. All the Thomason studies compared the two conditions, Pretest and Posttest, within participants, and therefore used the paired design. This dataset is from the first Thomason study, Thomason 1. IT used a group of N = 12 students, whose critical- thinking ability was assessed at Pretest and Posttest using the Logical Reasoning section of the Law School Aptitude Test (LSAT).

Usage

data_thomason1

Format

data_thomason1 A data frame with 12 rows and 2 columns:

pretest

LSAT score before argument mapping and mastery training

posttest

LSAT score after argument mapping and mastery training

Source

Thomason, N. R., Adajian, T., Barnett, A. E., Boucher, S., van der Brugge, E., Campbell, J., Knorpp, W., Lempert, R., Lengbeyer, L., Mandel, D. R., Rider, Y., van Gelder, T., & Wilkins, J. (2014). Critical thinking final report. The University of Melbourne, N66001-12-C-2004.