Estimates for a multi-group design with a continuous outcome variable

Returns object estimate_mdiff_ind_contrast is suitable for a multi-group design (between subjects) with a continuous outcome variable. It accepts a user-defined set of contrast weights that allows estimation of any 1-df contrast. It can express estimates as mean differences, standardized mean differences (Cohen's d) or median differences (raw data only). You can pass raw data or summary data.

Usage

estimate_mdiff_ind_contrast(
  data = NULL,
  outcome_variable = NULL,
  grouping_variable = NULL,
  means = NULL,
  sds = NULL,
  ns = NULL,
  contrast = NULL,
  grouping_variable_levels = NULL,
  outcome_variable_name = "My outcome variable",
  grouping_variable_name = "My grouping variable",
  conf_level = 0.95,
  assume_equal_variance = FALSE,
  save_raw_data = TRUE
)

Arguments

data

For raw data - a data frame or tibble

outcome_variable

For raw data - The column name of the outcome variable, or a vector of numeric data

grouping_variable

For raw data - The column name of the grouping variable, or a vector of group names

means

For summary data - A vector of 2 or more means

sds

For summary data - A vector of standard deviations, same length as means

ns

For summary data - A vector of sample sizes, same length as means

contrast

A vector of group weights, same length as number of groups.

grouping_variable_levels

For summary data - An optional vector of group labels, same length as means

outcome_variable_name

Optional friendly name for the outcome variable. Defaults to 'My outcome variable' or the outcome variable column name if a data frame is passed.

grouping_variable_name

Optional friendly name for the grouping variable. Defaults to 'My grouping variable' or the grouping variable column name if a data.frame is passed.

conf_level

The confidence level for the confidence interval. Given in decimal form. Defaults to 0.95.

assume_equal_variance

Defaults to FALSE

save_raw_data

For raw data; defaults to TRUE; set to FALSE to save memory by not returning raw data in estimate object

Value

Returns object of class esci_estimate

es_mean_difference

type -

outcome_variable_name -

grouping_variable_name -

effect -

effect_size -

LL -

UL -

SE -

df -

ta_LL -

ta_UL -

es_median_difference

type -

outcome_variable_name -

grouping_variable_name -

effect -

effect_size -

LL -

UL -

SE -

ta_LL -

ta_UL -

es_smd

outcome_variable_name -

grouping_variable_name -

effect -

effect_size -

LL -

UL -

numerator -

denominator -

SE -

df -

d_biased -

overview

outcome_variable_name -

grouping_variable_name -

grouping_variable_level -

mean -

mean_LL -

mean_UL -

median -

median_LL -

median_UL -

sd -

min -

max -

q1 -

q3 -

n -

missing -

df -

mean_SE -

median_SE -

raw_data

grouping_variable -

outcome_variable -

Details

Reach for this function in place of a one-way ANOVA.

Once you generate an estimate with this function, you can visualize it with plot_mdiff() and you can test hypotheses with test_mdiff().

The estimated mean differences are from statpsych::ci.lc.mean.bs().

The estimated SMDs are from CI_smd_ind_contrast() which relies on statpsych::ci.lc.stdmean.bs() unless there are only 2 groups.

The estimated median differences are from statpsych::ci.lc.median.bs()

Examples

# From summary data
data("data_halagappa")

estimate <- estimate_mdiff_ind_contrast(
  means = data_halagappa$Mean,
  sds = data_halagappa$SD,
  ns = data_halagappa$n,
  grouping_variable_levels = as.character(data_halagappa$Groups),
  assume_equal_variance = TRUE,
  contrast = c(
    "NFree10" = 1/3,
    "AFree10" = 1/3,
    "ADiet10" = -1/3,
    "NFree17" = -1/3,
    "AFree17" = 1/3,
    "ADiet17" = -1/3
  ),
  grouping_variable_name = "Diet",
  outcome_variable_name = "% time near target"
)
estimate
#> Analysis of raw data:
#> Outcome variable(s) = % time near target
#> Grouping variable(s) = Diet
#> 
#> ---Overview---
#>   outcome_variable_name grouping_variable_name grouping_variable_level mean
#> 1    % time near target                   Diet                 NFree10 37.5
#> 2    % time near target                   Diet                 AFree10 31.9
#> 3    % time near target                   Diet                 ADiet10 41.2
#> 4    % time near target                   Diet                 NFree17 33.4
#> 5    % time near target                   Diet                 AFree17 29.9
#> 6    % time near target                   Diet                 ADiet17 38.3
#>    mean_LL  mean_UL   sd  n  df  mean_SE
#> 1 32.32519 42.67481 10.0 19 108 2.610673
#> 2 26.72519 37.07481 13.5 19 108 2.610673
#> 3 36.02519 46.37481 14.8 19 108 2.610673
#> 4 28.22519 38.57481 10.0 19 108 2.610673
#> 5 24.72519 35.07481  8.7 19 108 2.610673
#> 6 33.12519 43.47481 10.0 19 108 2.610673
#> 
#> -- es_mean_difference --
#>         type outcome_variable_name grouping_variable_name
#> 1 Comparison    % time near target                   Diet
#> 2  Reference    % time near target                   Diet
#> 3 Difference    % time near target                   Diet
#>                                                                  effect
#> 1                                     (NFree10 and AFree10 and AFree17)
#> 2                                     (ADiet10 and NFree17 and ADiet17)
#> 3 (NFree10 and AFree10 and AFree17) ‒ (ADiet10 and NFree17 and ADiet17)
#>   effect_size        LL         UL       SE  df     ta_LL      ta_UL
#> 1   33.100000 30.112324 36.0876762 1.507273 108 30.599306 35.6006939
#> 2   37.633333 34.645657 40.6210095 1.507273 108 35.132639 40.1340273
#> 3   -4.533333 -8.758546 -0.3081211 2.131606 108 -8.069849 -0.9968181
#> 
#> -- es_smd --
#>   outcome_variable_name grouping_variable_name
#> 1    % time near target                   Diet
#>                                                                  effect
#> 1 (NFree10 and AFree10 and AFree17) ‒ (ADiet10 and NFree17 and ADiet17)
#>   effect_size         LL          UL numerator denominator        SE  df
#> 1  -0.3955976 -0.7655978 -0.02380328 -4.533333    11.37966 0.1892368 108
#>     d_biased
#> 1 -0.3983716
#> This standardized mean difference is called d_s because the standardizer used was s_p. d_s has been corrected for bias. Correction for bias can be important when df < 50.  See the rightmost column for the biased value.
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

if (FALSE) {
# To visualize the estimated mean difference
plot_mdiff(estimate, effect_size = "mean")
}