Estimate ameta-analytic Cohen's d1 across multiple studies

meta_d1 is suitable for synthesizing across multiple single-group studies with a continuous outcome variable, but where the outcome is not measured on the same scale in all studies

Usage

meta_d1(
  data,
  ds,
  ns,
  labels = NULL,
  moderator = NULL,
  contrast = NULL,
  effect_label = "My effect",
  random_effects = TRUE,
  conf_level = 0.95
)

Arguments

data

A data frame or tibble

ds

Set of bias-adjusted cohen's d1 values, 1 for each study

ns

Set of sample sizes, positive integers, 1 for each study

labels

Optional set of labels, 1 for each study

moderator

Optional factor as a categorical moderator; should have k > 2 per group

contrast

Optional vector specifying a contrast between moderator levels

effect_label

Optional character providing a human-friendly label for the effect

random_effects

Boolean; TRUE for a random effects model; otherwise fixed effects

conf_level

The confidence level for the confidence interval. Given in decimal form. Defaults to 0.95.

Value

An esci-estimate object; a list of data frames and properties. Returned tables include:

es_meta - A data frame with one overall. If a moderator was defined, there is an additional row for each level of the moderator.

effect_label - Study label

effect_size - Effect size

LL - Lower bound of conf_level% confidence interval

UL - Upper bound of conf_level% confidence interval

SE - Expected standard error

k - Number of studies

diamond_ratio - ratio of random to fixed effects meta-analtics effect sizes

diamond_ratio_LL - lower bound of conf_level% confidence interval for diamond ratio

diamond_ratio_UL - upper bound of conf_level% confidence interval for diamond ratio

I2 - I2 measure of heterogeneity

I2_LL - Lower bound of conf_level% confidence interval for I2

I2_UL - upper bound of conf_level% confidence interval for I2

PI_LL - lower bound of conf_level% of prediction interval

PI_UL - upper bound of conf_level% of prediction interval

p - p value for the meta-analytic effect size, based on null of exactly 0

width - width of the effect-size confidence interval

FE_effect_size - effect size of the fixed-effects model (regardless of if fixed effects was selected

RE_effect_size - effect size of the random-effects model (regardless of if random effects was selected

FE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio

RE_CI_width - width of the fixed-effects confidence interval, used to calculate diamond ratio

es_heterogeneity - A data frame of of heterogeneity values and conf_level% CIs for the meta-analytic effect size. If a moderator was defined also reports heterogeneity estimates for each level of the moderator.

effect_label - study label

moderator_variable_name - if moderator passed, gives name of the moderator

moderator_variable_name - 'Overall' and each level of moderator, if passed

measure - Name of the measure of heterogeneity

LL - lower bound of conf_level% confidence interval

UL - upper bound of conf_level% confidence interval

raw_data - A data from with one row for each study that was passed

label - study label

effect_size - effect size

weight - study weight in the meta analysis

sample_variance - expected level of sampling variation

SE - expected standard error

LL - lower bound of conf_level% confidence interval

UL - upper bound of conf_level% confidence interval

mean - used to calculate study p value; this is the d value entered for the study

n - study sample size

sd - use to calculate study p value; set to 1 for each study

p - p value for the study, based on null of exactly 0

Details

Each study's effect size should be expressed as Cohen's d1: (mean - reference) / sd.

And the d1 values should all be corrected for bias. The function CI_smd_one() can assist with converting raw data from each study to d1_unbiased.

The meta-analytic effect size, confidence interval and heterogeneity estimates all come from metafor::rma().

The diamond ratio and its confidence interval come from CI_diamond_ratio().