Plot an estimated Pearson's r value

plot_correlation creates a ggplot2 plot suitable for visualizing an estimate correlation between two continuous variables (Pearson's r).

Usage

plot_correlation(
  estimate,
  error_layout = c("halfeye", "eye", "gradient", "none"),
  error_scale = 0.3,
  error_normalize = c("groups", "all", "panels"),
  rope = c(NA, NA),
  ggtheme = NULL
)

Arguments

estimate

An esci_estimate object generated by estimate_magnitude()

error_layout

Optional; One of 'halfeye', 'eye', 'gradient' or 'none' for how expected sampling error of the measure of central tendency should be displayed. Caution - the displayed error distributions do not seem correct yet

error_scale

Optional real number > 0 speciyin width of the expected sampling error visualization; default is 0.3

error_normalize

Optional; One of 'groups' (default), 'all', or 'panels' specifying how width of expected sampling error distributions should be calculated.

rope

Optional two-item vector specifying a region of practical equivalence (ROPE) to be highlighted on the plot. For a point null hypothesis, pass the same value (e.g. c(0, 0) to test a point null of exaclty 0); for an interval null pass ascending values (e.g. c(-1, 1))

ggtheme

Optional ggplot2 theme object to control overall styling; defaults to ggplot2::theme_classic()

Details

This function can be passed an esci_estimate object generated by [esci::estimate_correlation())]

[esci::estimate_correlation())]: R:esci::estimate_correlation())