Plots for comparing Pearson r values between conditions

plot_rdiff helps visualize comparisons of Pearson's r estimates between conditions. It plots the Pearson's r value for each level of a grouping variable and emphasizes a 1-df comparison among conditions, plotting the estimated difference and its confidence interval with a difference axis. You can pass esci-estimate objects generated by estimate_rdiff_two(). This function returns a ggplot2 object.

Usage

plot_rdiff(
  estimate,
  error_layout = c("halfeye", "eye", "gradient", "none"),
  error_scale = 0.3,
  error_normalize = c("groups", "all", "panels"),
  difference_axis_breaks = 5,
  simple_contrast_labels = TRUE,
  ylim = c(NA, NA),
  ybreaks = 5,
  rope = c(NA, NA),
  ggtheme = NULL
)

Arguments

estimate

An esci-estimate object generated by an estimate_pdiff_ function

error_layout

Optional; one of 'halfeye', 'eye', 'gradient' or 'none' to determine how expected error distribution will be displayed for each estimated parameter. Defaults to 'halfeye'. Currently does not apply if 'median' is selected as effect size, in which case a simple error bar will be used

error_scale

Optional numeric determining width of the expected error distribution. Defaults to 0.3

error_normalize

Optional; one of 'groups', 'all', or 'panels' to determine how width of the expected error distributions will be normalized. Defaults to 'groups'. See documentation in ggdist

difference_axis_breaks

Optional numeric > 1 of suggested number of breaks for the difference axis. Defaults to 5

simple_contrast_labels

Optional logical to determine if contrasts are given simple labels ('Reference', 'Comparison', 'Difference') or more descriptive labels based on the contrast specified.

ylim

Optional 2-item vector specifying y-axis limits. Defaults to c(NA NA); Use NA to specify auto-limit.

ybreaks

Optional numeric > 2 for suggested number of y-axis breaks; defaults to 5

rope

Optional 2-item vector with item 2 >= item 1. Use to specify a range of values to use to visualize a hypothesis test. If both values are the same, a point-null hypothesis test will be visualized. If item2 > item1 an interval-null hypothesis test will be visualized. Defaults to c(NA, NA), which is to not visualize a hypothesis test

ggtheme

Optional ggplot2 theme object to specify the visual style of the plot. Defaults to ggplot2::theme_classic()

Value

Returns a ggplot object

Details

This function was developed primarily for student use within jamovi when learning along with the text book Introduction to the New Statistics, 2nd edition (Cumming & Calin-Jageman, 2024).

Expect breaking changes as this function is improved for general use. Work still do be done includes:

Revise to avoid deprecated ggplot features

Revise for consistent ability to control aesthetics and consistent layer names

Examples

# From summary data
estimate <- esci::estimate_rdiff_two(
  comparison_r = .53,
  comparison_n = 45,
  reference_r = .41,
  reference_n = 59,
  grouping_variable_levels = c("Females", "Males"),
  x_variable_name = "Satisfaction with life",
  y_variable_name = "Body satisfaction",
  grouping_variable_name = "Gender",
  conf_level = .95
)
estimate
#> Analysis of summary data:
#> 
#> -- es_r --
#>   grouping_variable_name grouping_variable_level        x_variable_name
#> 1                 Gender                 Females Satisfaction with life
#> 2                 Gender                   Males Satisfaction with life
#>     y_variable_name  effect effect_size        LL        UL        SE  n df
#> 1 Body satisfaction Females        0.41 0.1685419 0.6005378 0.1092338 59 57
#> 2 Body satisfaction   Males        0.53 0.2744717 0.7096862 0.1084084 45 43
#>       ta_LL     ta_UL
#> 1 0.2091420 0.5729339
#> 2 0.3188047 0.6847105
#> 
#> -- es_r_difference --
#>         type grouping_variable_name grouping_variable_level
#> 1 Comparison                 Gender                   Males
#> 2  Reference                 Gender                 Females
#> 3 Difference                 Gender       Males  -  Females
#>          x_variable_name   y_variable_name            effect effect_size
#> 1 Satisfaction with life Body satisfaction             Males        0.53
#> 2 Satisfaction with life Body satisfaction           Females        0.41
#> 3 Satisfaction with life Body satisfaction Males  -  Females        0.12
#>           LL        UL        SE  n df      ta_LL     ta_UL        rz       sem
#> 1  0.2744717 0.7096862 0.1084084 45 43  0.3188047 0.6847105 0.5901452 0.1543033
#> 2  0.1685419 0.6005378 0.1092338 59 57  0.2091420 0.5729339 0.4356112 0.1336306
#> 3 -0.1987466 0.4209803        NA NA NA -0.1467413 0.3735336 0.1545339 0.2041241
#>           z            p
#> 1 3.8245778 0.0001309964
#> 2 3.2598159 0.0011148455
#> 3 0.7570586 0.4490147644
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

# To evaluate a hypothesis (interval null from -0.1 to 0.1):
esci::test_rdiff(estimate)
#> $properties
#> $properties$effect_size_name
#> [1] "r"
#> 
#> $properties$alpha
#> [1] 0.05
#> 
#> $properties$interval_null
#> [1] FALSE
#> 
#> $properties$rope
#> [1] 0 0
#> 
#> $properties$rope_units
#> [1] "raw"
#> 
#> 
#> $point_null
#>             test_type                        outcome_variable_name
#> 1 Nil Hypothesis Test Satisfaction with life and Body satisfaction
#>              effect null_words confidence         LL        UL
#> 1 Males  -  Females       0.00         95 -0.1987466 0.4209803
#>                               CI              CI_compare         t df         p
#> 1 95% CI [-0.1987466, 0.4209803] The 95% CI contains H_0 0.7570586 NA 0.4490148
#>   p_result      null_decision
#> 1 p ≥ 0.05 Fail to reject H_0
#>                                              conclusion significant
#> 1 At α = 0.05, 0.00 remains a plausible value of Ͱ_diff       FALSE
#> 

if (FALSE) {
# To visualize the values of r and their difference
esci::plot_rdiff(estimate)
}