Estimate the diamond ratio for a meta-analtyic effect, a measure of heterogeneity

CI_diamond_ratio returns the diamond ratio and CI for a meta-analtyic effect, the ratio of the random-effects CI width to the fixed-effects CI width. The diamond ratio is a measure of effect-size heteogeneity.

Usage

CI_diamond_ratio(RE, FE, vi, conf_level = 0.95)

Source

Cairns, Maxwell, Geoff Cumming, Robert Calin‐Jageman, and Luke A. Prendergast. “The Diamond Ratio: A Visual Indicator of the Extent of Heterogeneity in Meta‐analysis.” British Journal of Mathematical and Statistical Psychology 75, no. 2 (May 2022): 201–19. https://doi.org/10.1111/bmsp.12258.

Arguments

RE

metafor object with random effects result

FE

metafor object with fixed effects result

vi

vector of effect size variances

conf_level

The confidence level for the confidence interval. Given in decimal form. Defaults to 0.95.

Value

Returns a list with 3 properties:

diamond_ratio

LL - lower limit of the conf_level% CI

UL - upper limit of the conf_level% CI

Details

Calculation of the CI is based on code provided by Maxwell Cairns (see Cairns et al., 2022). Specifically, this function implements what Cairns et al (2022) called the bWT-DL approach, which had the best peformance for a diamond ratio CI.