Macnamara r ma - Ch11 - from Macnamara et al. (2014)

Is genius born or made? Could any of us be Michael Jordan or Mozart if we worked sufficiently hard to develop the requisite skills? Meta-analysis of correlations can help answer such questions. The issue here is the extent that practice and effort may be sufficient for achieving the highest levels of expertise. Ericsson et al. (1993) argued that years of effort is what matters most: 'Many characteristics once believed to reflect innate talent are actually the result of intense practice extended for a minimum of 10 years' (p. 363). This view was enormously popularized by Malcolm Gladwell (2008), who argued in his book Outliers that 10,000 hours of focused practice is the key to achieving expertise. However, this view is now being challenged, with one important contribution being a large meta-analysis of correlations between amount of intense practice and level of achievement: Macnamara et al. (2014) combined 157 correlations reported in a wide range of fields, from sports to music and education, and found correlation of only r = .35 (.30, .39). Table 11.1 shows the 16 main correlations for music, from Macnamara et al. (2014).

Usage

data_macnamara_r_ma

Format

data_macnamara_r_ma

A data frame with 16 rows and 4 columns:

Study

factor

r

numeric

N

integer

Instrument_Type

factor

Source

doi:10.1177/0956797614535810