PowerPerformance ma - Ch9 - from Burgmer and Englich (2012), and Cusack et al. (2015)

To what extent could feeling powerful affect your performance at motor skills? To investigate, Burgmer and Englich (2012) assigned German participants to either power or control conditions and then asked them to play golf (Experiment 1) or darts (Experiment 2). They found that participants manipulated to feel powerful performed substantially better than those in the control condition. To study this finding further, Cusack et al. (2015) conducted five replications in the United States. Across these replications they tried different ways of manipulating power, different types of tasks (golf, mirror tracing, and a cognitive task), different levels of difficulty, and different types of participant pools (undergraduates and online). Summary data from all seven studies are available in PowerPerformance ma.

Usage

data_powerperformance_ma

Format

data_powerperformance_ma

A data frame with 8 rows and 12 columns:

StudyName

factor

Country

factor

Population

factor

Difficulty

factor

Task

factor

M Control

numeric

s Control

numeric

M Power

numeric

s Power

numeric

Cohen d unb

numeric

n Control

integer

n Power

integer

Source

doi:10.1371/journal.pone.0140806