Estimates for a single-group design with a continuous outcome variable compared to a reference or population value

estimate_mdiff_one is suitable for a single-group design with a continuous outcome variable that is compared to a reference or population value. It can express estimates as mean differences, standardized mean differences (Cohen's d) or median differences (raw data only). You can pass raw data or summary data.

Usage

estimate_mdiff_one(
  data = NULL,
  outcome_variable = NULL,
  comparison_mean = NULL,
  comparison_sd = NULL,
  comparison_n = NULL,
  reference_mean = 0,
  outcome_variable_name = "My outcome variable",
  conf_level = 0.95,
  save_raw_data = TRUE
)

Arguments

data

For raw data - a data frame or tibble

outcome_variable

For raw data - The column name of the outcome variable, or a vector of numeric data

comparison_mean

For summary data, a numeric

comparison_sd

For summary data, numeric > 0

comparison_n

For summary data, a numeric integer > 0

reference_mean

Reference value, defaults to 0

outcome_variable_name

Optional friendly name for the outcome variable. Defaults to 'My outcome variable' or the outcome variable column name if a data frame is passed.

conf_level

The confidence level for the confidence interval. Given in decimal form. Defaults to 0.95.

save_raw_data

For raw data; defaults to TRUE; set to FALSE to save memory by not returning raw data in estimate object

Value

Returns object of class esci_estimate

overview

outcome_variable_name -

mean -

mean_LL -

mean_UL -

median -

median_LL -

median_UL -

sd -

min -

max -

q1 -

q3 -

n -

missing -

df -

mean_SE -

median_SE -

es_mean

outcome_variable_name -

effect -

effect_size -

LL -

UL -

SE -

df -

ta_LL -

ta_UL -

es_median

outcome_variable_name -

effect -

effect_size -

LL -

UL -

SE -

df -

ta_LL -

ta_UL -

raw_data

grouping_variable -

outcome_variable -

es_mean_difference

outcome_variable_name -

effect -

effect_size -

LL -

UL -

SE -

df -

ta_LL -

ta_UL -

type -

es_median_difference

outcome_variable_name -

effect -

effect_size -

LL -

UL -

SE -

df -

ta_LL -

ta_UL -

type -

es_smd

outcome_variable_name -

effect -

effect_size -

LL -

UL -

numerator -

denominator -

SE -

df -

d_biased -

Details

Reach for this function in place of a z-test or one-sample t-test.

Once you generate an estimate with this function, you can visualize it with plot_mdiff() and you can test hypotheses with test_mdiff().

The estimated mean differences are from statpsych::ci.mean1().

The estimated SMDs are from CI_smd_one().

The estimated median differences are from statpsych::ci.median1()

Examples

# From raw data
data("data_penlaptop1")
estimate <- esci::estimate_mdiff_one(
  data = data_penlaptop1[data_penlaptop1$condition == "Pen", ],
  outcome_variable = transcription,
  reference_mean = 10
)
estimate
#> Analysis of raw data:
#> Data frame = data
#> Outcome variable(s) = transcription
#> 
#> ---Overview---
#>   outcome_variable_name     mean  mean_LL  mean_UL median median_LL median_UL
#> 1         transcription 8.811765 7.154642 10.46889    8.6       6.5      11.1
#>         sd min  max  q1     q3  n missing df   mean_SE median_SE
#> 1 4.749339   1 20.1 5.2 11.275 34       0 33 0.8145049  1.021201
#> 
#> -- es_mean --
#>   outcome_variable_name        effect effect_size       LL       UL        SE
#> 1         transcription transcription    8.811765 7.154642 10.46889 0.8145049
#>   df    ta_LL    ta_UL
#> 1 33 7.746606 9.876923
#> 
#> -- es_median --
#>   outcome_variable_name        effect effect_size  LL   UL       SE df ta_LL
#> 1         transcription transcription         8.6 6.5 11.1 1.021201 33   7.1
#>   ta_UL
#> 1  10.8
#> 
#> -- es_mean_difference --
#>   outcome_variable_name                          effect effect_size        LL
#> 1         transcription                   transcription    8.811765  7.154642
#> 2         transcription                 Reference value   10.000000        NA
#> 3         transcription transcription ‒ Reference value   -1.188235 -2.845358
#>           UL        SE df     ta_LL     ta_UL       type
#> 1 10.4688874 0.8145049 33  7.746606  9.876923 Comparison
#> 2         NA        NA NA  7.746606  9.876923  Reference
#> 3  0.4688874 0.8145049 33 -2.253394 -0.123077 Difference
#> 
#> -- es_median_difference --
#>   outcome_variable_name                          effect effect_size   LL   UL
#> 1         transcription                   transcription         8.6  6.5 11.1
#> 2         transcription                 Reference value        10.0   NA   NA
#> 3         transcription transcription ‒ Reference value        -1.4 -3.5  1.1
#>         SE df ta_LL ta_UL       type
#> 1 1.021201 33   7.1  10.8 Comparison
#> 2       NA NA    NA    NA  Reference
#> 3 1.021201 33  -2.9   0.8 Difference
#> 
#> -- es_smd --
#>   outcome_variable_name                          effect effect_size         LL
#> 1         transcription transcription ‒ Reference value  -0.2444527 -0.5838873
#>           UL numerator denominator        SE df   d_biased
#> 1 0.09856549 -1.188235    4.749339 0.1740983 33 -0.2501896
#> This standardized mean difference is called d_1 because the standardizer used was s. d_1 has been corrected for bias. Correction for bias can be important when df < 50.  See the rightmost column for the biased value.
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

if (FALSE) {
# To visualize the estimate
plot_mdiff(estimate)
}

if (FALSE) {
# To conduct a hypothesis test
test_mdiff(estimate, effect_size = "mean", rope = c(-2, 2))
}

# From summary data
mymean <- 12.09
mysd <- 5.52
myn <- 103

estimate <- esci::estimate_mdiff_one(
  comparison_mean = mymean,
  comparison_sd = mysd,
  comparison_n = myn,
  reference_mean = 12
)
estimate
#> Analysis of raw data:
#> Outcome variable(s) = My outcome variable
#> 
#> ---Overview---
#>   outcome_variable_name  mean  mean_LL  mean_UL   sd   n  df   mean_SE
#> 1   My outcome variable 12.09 11.01117 13.16883 5.52 103 102 0.5439018
#> 
#> -- es_mean --
#>   outcome_variable_name              effect effect_size       LL       UL
#> 1   My outcome variable My outcome variable       12.09 11.01117 13.16883
#>          SE  df    ta_LL    ta_UL
#> 1 0.5439018 102 11.38842 12.79158
#> 
#> -- es_mean_difference --
#>   outcome_variable_name                                effect effect_size
#> 1   My outcome variable                   My outcome variable       12.09
#> 2   My outcome variable                       Reference value       12.00
#> 3   My outcome variable My outcome variable ‒ Reference value        0.09
#>           LL        UL        SE  df      ta_LL      ta_UL       type
#> 1 11.0111734 13.168827 0.5439018 102 11.3884176 12.7915824 Comparison
#> 2         NA        NA        NA  NA 11.3884176 12.7915824  Reference
#> 3 -0.9888266  1.168827 0.5439018 102 -0.6115824  0.7915824 Difference
#> 
#> -- es_smd --
#>   outcome_variable_name                                effect effect_size
#> 1   My outcome variable My outcome variable ‒ Reference value  0.01618412
#>           LL        UL numerator denominator         SE  df   d_biased
#> 1 -0.1769892 0.2092782      0.09        5.52 0.09853943 102 0.01630435
#> This standardized mean difference is called d_1 because the standardizer used was s. d_1 has been corrected for bias. Correction for bias can be important when df < 50.  See the rightmost column for the biased value.
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

if (FALSE) {
# To visualize the estimate
plot_mdiff(estimate)
}

if (FALSE) {
# To conduct a hypothesis test
test_mdiff(estimate, effect_size = "mean", rope = c(-2, 2))
}