Generates a forest plot displaying results of a meta-analysis

`plot_meta' returns a ggplot2 object visualizing the results of a meta-analysis, showing each study effect size and CI, the overall effect size and CI as a diamond, effect sizes estimated at each moderator level (if defined), and (optionally) prediction intervals for subsequent studies. This function requires as input an esci_estimate object generated by an esci meta-analysis function: meta_any(), meta_d1(), meta_d2(), meta_mdiff_two(), meta_mean(), meta_pdiff_two(), meta_proportion(), and meta_r().

Usage

plot_meta(
  estimate,
  mark_zero = TRUE,
  include_PIs = FALSE,
  report_CIs = FALSE,
  explain_DR = FALSE,
  meta_diamond_height = 0.35,
  ggtheme = ggplot2::theme_classic()
)

Arguments

estimate

an esci_estimate object generated by an esci meta_ function

mark_zero

Boolean; defaults to TRUE to include a dotted line indicated no effect (effect_size = 0)

include_PIs

Boolean; defaults to FALSE; set to TRUE to include prediction intervals for the overall effect and each moderator level (if defined)

report_CIs

Boolean; defaults to FALSE; set to TRUE to include printed representation of each study effect size and CI along the right- hand of the figure

explain_DR

Boolean; dedaults to FALSE; set to TRUE if no moderator is defined to show both the RE and FE effect sizes to represent how the diamond ration measure of effect-size heterogeneity is calculated

meta_diamond_height

Optional real number > 0 to indicate height that each meta-analytic diamond should be drawn; defaults to 0.35

ggtheme

Optional ggplot2 theme object to control overall styling; defaults to ggplot2::theme_classic()

Value

Returns a ggplot object

Details

This function was developed primarily for student use within jamovi when learning along with the text book Introduction to the New Statistics, 2nd edition (Cumming & Calin-Jageman, 2024).

Expect breaking changes as this function is improved for general use. Work still do be done includes:

Revise to avoid deprecated ggplot features

Revise for consistent ability to control aesthetics and consistent layer names

Examples

# Data set -- see Introduction to the New Statistics, 1st edition
brain_scan_persuasion <- data.frame(
  study_name = c("McCabe 1", "McCabe 2", paste("Michael",
     seq(1:10), sep = " ")),
  nbM = c(2.89, 2.69, 2.90, 2.62, 2.96, 2.93, 2.86,
     2.50, 2.41, 2.54, 2.73, 2.66),
  nbS = c(0.79, 0.55, 0.58, 0.54, 0.36, 0.60, 0.59,
     0.84, 0.78, 0.66, 0.67, 0.65),
  nbN = c(28, 26, 98, 42, 24, 184, 274, 58, 34, 99,
     98, 94),
  bM = c(3.12, 3.00, 2.86, 2.85, 3.07, 2.89, 2.91,
     2.60, 2.74, 2.72, 2.68, 2.64),
  bS = c(0.65, 0.54, 0.61, 0.57, 0.55, 0.60, 0.52,
     0.83, 0.51, 0.68, 0.69, 0.71),
  bN = c(26, 28, 99, 33, 21, 184, 255, 55, 34, 95,
     93, 97),
  mod = as.factor(
    c("Simple", "Critique", "Simple","Simple","Simple",
       "Simple","Simple","Critique","Critique","Critique",
       "Critique","Critique")
  ),
  ds = c(0.31217944,  0.56073138, -0.06693802,  0.41136192,
     0.23581389, -0.06652995,  0.08958082,  0.11892778,
     0.49506069,  0.26765910, -0.07326, -0.02925)
)

# Meta-analysis: random effects, no moderator
estimate <- esci::meta_mdiff_two(
  data = brain_scan_persuasion,
  comparison_means = bM,
  comparison_sds = bS,
  comparison_ns = bN,
  reference_means = nbM,
  reference_sds = nbS,
  reference_ns = nbN,
  labels = study_name,
  effect_label = "Brain Photo Rating - No Brain Photo Rating",
  assume_equal_variance = TRUE,
  random_effects = TRUE
)
if (FALSE) {
# Forest plot
esci::plot_meta(estimate)
}

# Meta-analysis: random effects, moderator
if (FALSE) {
# Forest plot
esci::plot_meta(estimate)
}

# Meta-analysis: random effects, moderator, output d_s
if (FALSE) {
# Forest plot
esci::plot_meta(estimate)
}