Generates a scatter plot of data for two continuous variables

plot_scatter returns a ggplot2 object of data from two continuous variables. Can indicate regression line and its confidence interval,prediction intervals regression residuals and more. This function requires as input an esci_estimate object generated by estimate_r()

Usage

plot_scatter(
  estimate,
  show_line = FALSE,
  show_line_CI = FALSE,
  show_PI = FALSE,
  show_residuals = FALSE,
  show_mean_lines = FALSE,
  show_r = FALSE,
  predict_from_x = NULL,
  plot_as_z = FALSE,
  ggtheme = ggplot2::theme_classic()
)

Arguments

estimate

an esci_estimate object generated by estimate_r()

show_line

Boolean; defaults to FALSE; set to TRUE to show the regression line

show_line_CI

Boolean; defaults to FALSE; set to TRUE to show the confidence interval on the regression line

show_PI

Boolean; defaults to FALSE; set to TRUE to show prediction intervals

show_residuals

Boolean; defaults to FALSE; set to TRUE to show residuals of prediction

show_mean_lines

Boolean; defaults to FALSE; set to TRUE to plot lines showing the mean of each variable

show_r

Boolean; defaults to FALSE; set to TRUE to print the r value and its CI on the plot

predict_from_x

Optional real number in the range of the x variable for the plot; Defaults to NULL; if passed, the graph shows the predicted Y' for this x value

plot_as_z

Boolean; defaults to FALSE; set to TRUE to convert x and y scores to z scores prior to plotting

ggtheme

Optional ggplot2 theme object to control overall styling; defaults to ggplot2::theme_classic()

Value

Returns a ggplot object

Details

This function was developed primarily for student use within jamovi when learning along with the text book Introduction to the New Statistics, 2nd edition (Cumming & Calin-Jageman, 2024).

Expect breaking changes as this function is improved for general use. Work still do be done includes:

Revise to avoid deprecated ggplot features

Revise for consistent ability to control aesthetics and consistent layer names

Examples

# From raw data

thomason1 <- data.frame(
  ls_pre = c(
    13,
    12,
    12,
    9,
    14,
    17,
    14,
    9,
    6,
    7,
    11,
    15
  ),
  ls_post = c(
    14,
    13,
    16,
    12,
    15,
    18,
    13,
    10,
    10,
    8,
    14,
    16
  )
)

estimate <- esci::estimate_r(
  thomason1,
  ls_pre,
  ls_post
)
estimate
#> Analysis of raw data:
#> Data frame = data
#> 
#> ---Overview---
#>   outcome_variable_name     mean   mean_LL  mean_UL median median_LL median_UL
#> 1                ls_pre 11.58333  9.476776 13.68989   12.0         9        14
#> 2               ls_post 13.25000 11.410019 15.08998   13.5        10        16
#>         sd min max   q1    q3  n missing df   mean_SE median_SE
#> 1 3.315483   6  17  9.0 14.00 12       0 11 0.9570974  1.208488
#> 2 2.895922   8  18 11.5 15.25 12       0 11 0.8359806  1.450186
#> 
#> -- es_r --
#>   x_variable_name y_variable_name             effect effect_size      LL
#> 1          ls_pre         ls_post ls_pre and ls_post   0.8923908 0.62894
#>         UL         SE  n df     ta_LL     ta_UL
#> 1 0.967157 0.06139936 12 10 0.6882891 0.9596343
#> 
#> -- regression --
#>       component    values        LL       UL
#> 1 Intercept (a) 4.2212267 0.8856544 7.556799
#> 2     Slope (b) 0.7794624 0.5017391 1.057186
#> [1] "Ŷ = 4.221 + 0.7795*X"
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

# To evaluate a hypothesis (interval null from -0.1 to 0.1):
test_correlation(estimate, rope = c(-0.1, 0.1))
#> $properties
#> $properties$effect_size_name
#> [1] "r"
#> 
#> $properties$alpha
#> [1] 0.05
#> 
#> $properties$interval_null
#> [1] TRUE
#> 
#> $properties$rope
#> [1] -0.1  0.1
#> 
#> $properties$rope_units
#> [1] "raw"
#> 
#> 
#> $point_null
#>             test_type outcome_variable_name             effect null_words
#> 1 Nil Hypothesis Test    ls_pre and ls_post ls_pre and ls_post       0.00
#>   confidence      LL       UL                         CI
#> 1         95 0.62894 0.967157 95% CI [0.62894, 0.967157]
#>                        CI_compare        t df            p p_result
#> 1 The 95% CI does not contain H_0 4.300635 10 1.703091e-05 p < 0.05
#>   null_decision                                      conclusion significant
#> 1    Reject H_0 At α = 0.05, 0.00 is not a plausible value of Ͱ        TRUE
#> 
#> $interval_null
#>                     test_type outcome_variable_name             effect
#> 1 Practical significance test    ls_pre and ls_post ls_pre and ls_post
#>            rope confidence
#> 1 (-0.10, 0.10)         95
#>                                                          CI
#> 1 95% CI [0.62894, 0.967157]\n90% CI [0.6882891, 0.9596343]
#>               rope_compare p_result                             conclusion
#> 1 95% CI fully outside H_0 p < 0.05 At α = 0.05, conclude Ͱ is substantive
#>   significant
#> 1        TRUE
#> 

if (FALSE) {
# To visualize the value of r
plot_correlation(estimate)

# To visualize the data (scatterplot)
plot_scatter(estimate)

# To visualize the data (scatterplot) and use regression to obtain Y' from X
plot_scatter(estimate, predict_from_x = 10)

}

# From summary data
estimate <- esci::estimate_r(r = 0.536, n = 50)
estimate
#> Analysis of summary data:
#> 
#> -- es_r --
#>   x_variable_name y_variable_name                          effect effect_size
#> 1   My x variable   My y variable My x variable and My y variable       0.536
#>          LL        UL        SE  n df     ta_LL     ta_UL
#> 1 0.2978573 0.7058914 0.1018149 50 48 0.3391487 0.6820747
#> 
#> 
#> Note: LL and UL are lower and upper boundaries of confidence intervals with 95% expected coverage.

if (FALSE) {
# To visualize the estimate
plot_correlation(estimate)
}