Test a hypothesis about the strength of a Pearson's r correlation

test_correlation is suitable for testing a hypothesis about a the strength of correlation between two continuous variables (designs in which Pearson's r is a suitable measure of correlation).

Usage

test_correlation(estimate, rope = c(0, 0), output_html = FALSE)

Arguments

estimate

An esci_estimate object generated by the estimate_r function

rope

A two-element vector defining the Region of Practical Equivalence (ROPE). Specify c(0, 0) to test a point null of exactly 0. Specify any two ascending values between -1 and 1 to test an interval null (e.g. c(.25, .45) to test the hypothesis that Pearson's r in the population (rho) is between .25 and .45).

output_html

TRUE to return results in HTML; FALSE (default) to return standard output

Value

Returns a list with 1-2 data frames

point_null - always returned

test_type - 'Nil hypothesis test', meaning a test against H0 = 0

outcome_variable_name - Name of the outcome variable

effect - Label for the effect being tested

null_words - Express the null in words

confidence - Confidence level, integer (95 for 95%, etc.)

LL - Lower boundary of the confidence% CI for the effect

UL - Upper boundary of the confidence% CI for the effect

CI - Character representation of the CI for the effect

CI_compare - Text description of relation between CI and null

t - If applicable, t value for hypothesis test

df - If applicable, degrees of freedom for hypothesis test

p - If applicable, p value for hypothesis test

p_result - Text representation of p value obtained

null_decision - Text represention of the decision for the null

conclusion - Text representation of conclusion to draw

significant - TRUE/FALSE if significant at alpha = 1-CI

interval_null - returned only if an interval null is specified

test_type - 'Practical significance test', meaning a test against an interval null

outcome_variable_name -

effect - Name of the outcome variable

rope - Test representation of null interval

confidence - Confidence level, integer (95 for 95%, etc.)

CI - Character representation of the CI for the effect

rope_compare - Text description of relation between CI and null interval

p_result - Text representation of p value obtained

conclusion - Text representation of conclusion to draw

significant - TRUE/FALSE if significant at alpha = 1-CI

Details

This function can be passed an esci_estimate object generated by estimate_r().

It can test hypotheses about a specific value for the difference (a point null) or about a range of values (an interval null)

Examples

# example code
estimate <- esci::estimate_r(r = 0.536, n = 50)

# Test against a point null of exactly 0
test_correlation(estimate)
#> $properties
#> $properties$effect_size_name
#> [1] "r"
#> 
#> $properties$alpha
#> [1] 0.05
#> 
#> $properties$interval_null
#> [1] FALSE
#> 
#> $properties$rope
#> [1] 0 0
#> 
#> $properties$rope_units
#> [1] "raw"
#> 
#> 
#> $point_null
#>             test_type           outcome_variable_name
#> 1 Nil Hypothesis Test My x variable and My y variable
#>                            effect null_words confidence        LL        UL
#> 1 My x variable and My y variable       0.00         95 0.2978573 0.7058914
#>                              CI                      CI_compare        t df
#> 1 95% CI [0.2978573, 0.7058914] The 95% CI does not contain H_0 4.103289 48
#>              p p_result null_decision
#> 1 4.073183e-05 p < 0.05    Reject H_0
#>                                        conclusion significant
#> 1 At α = 0.05, 0.00 is not a plausible value of Ͱ        TRUE
#> 

# Test against an interval null (-0.1, 0.1)
test_correlation(estimate, rope = c(-0.1, 0.1))
#> $properties
#> $properties$effect_size_name
#> [1] "r"
#> 
#> $properties$alpha
#> [1] 0.05
#> 
#> $properties$interval_null
#> [1] TRUE
#> 
#> $properties$rope
#> [1] -0.1  0.1
#> 
#> $properties$rope_units
#> [1] "raw"
#> 
#> 
#> $point_null
#>             test_type           outcome_variable_name
#> 1 Nil Hypothesis Test My x variable and My y variable
#>                            effect null_words confidence        LL        UL
#> 1 My x variable and My y variable       0.00         95 0.2978573 0.7058914
#>                              CI                      CI_compare        t df
#> 1 95% CI [0.2978573, 0.7058914] The 95% CI does not contain H_0 4.103289 48
#>              p p_result null_decision
#> 1 4.073183e-05 p < 0.05    Reject H_0
#>                                        conclusion significant
#> 1 At α = 0.05, 0.00 is not a plausible value of Ͱ        TRUE
#> 
#> $interval_null
#>                     test_type           outcome_variable_name
#> 1 Practical significance test My x variable and My y variable
#>                            effect          rope confidence
#> 1 My x variable and My y variable (-0.10, 0.10)         95
#>                                                             CI
#> 1 95% CI [0.2978573, 0.7058914]\n90% CI [0.3391487, 0.6820747]
#>               rope_compare p_result                             conclusion
#> 1 95% CI fully outside H_0 p < 0.05 At α = 0.05, conclude Ͱ is substantive
#>   significant
#> 1        TRUE
#>